+++ CHANGES.md
@@ -4,6 +4,9 @@
   * Develop utility function to clone Entity instance ([#967](https://github.com/CiscoDevNet/ydk-gen/issues/967))
   * Go CodecService fails to encode leaf-list data ([#968](https://github.com/CiscoDevNet/ydk-gen/issues/968))
 
+#### Documentation improvements
+  * Bundle profile description is not accurate ([#971](https://github.com/CiscoDevNet/ydk-gen/issues/971))
+
 
 ### 2019-10-15 version 0.8.4