+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 #### Resolved GitHub issues
   * gNMI set RPC fails when JSON payload is formatted string
+  * Enhance generator to indicate how a YANG modules was processed ([#894](https://github.com/CiscoDevNet/ydk-gen/issues/894))
+  * YDK fails create bundle package when package name contains '.' ([#978](https://github.com/CiscoDevNet/ydk-gen/issues/978))
 
 #### Documentation improvements
   * Bundle profile description is not accurate ([#971](https://github.com/CiscoDevNet/ydk-gen/issues/971))